requirements.txt
requests
beautifulsoup4
boto3
pandas
aws-cdk-lib
constructs